fix: moved init from main to system_init
