+++ gripper_17.X/include/system_init.h
@@ -1,30 +1,86 @@
-/* 
- * File:   other.h
- * Author: nathaniel
- *
- * Created on January 19, 2025, 4:29 PM
- */
-
 #ifndef SYSTEM_INIT_H
-#define	SYSTEM_INIT_H
+#define SYSTEM_INIT_H
 
-#ifdef	__cplusplus
+#include <stdbool.h>
+#include <stddef.h>
+#include <stdint.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include "adc0.h"
+#include "can1.h"
+#include "can_common.h"
+#include "clock.h"
+#include "dma.h"
+#include "i2c.h"  // I2C client backup
+#include "pm.h"
+#include "rtc.h"
+#include "sercom1_i2c.h"  // Encoder i2c
+#include "tcc.h"
+#include "tcc0.h"
+#include "tcc_common.h"
+#include "usart.h"
+#include "wdt.h"
+
+// Encoder
+#define SHOULDER_ADDR 0x40
+#define WRIST_ADDR 0x41
+#define GRIP_ADDR 0x42
+#define ANGLE_REGISTER 0xFE
+#define I2C_TIMEOUT 100000
+#define NUM_ENCODERS 3
+
+#define TRANSFER_SIZE 16
+#define ADC_VREF 5.0f
+#define CURRENT_TRESHOLD 2.7f  // 1 A
+#define VOLTAGE_THRESHOLD 2048
+#define RTC_COMPARE_VAL 50
+
+#ifdef __cplusplus
 extern "C" {
 #endif
 
-void NVIC_Initialize(void); 
+typedef enum {
+    STATE_IDLE,
+    STATE_GRIPPER_ACTIVE,
+} STATES;
+
+typedef enum {
+    STATE_CAN_RECEIVE,
+    STATE_CAN_TRANSMIT,
+} CAN_STATES;
+
+typedef enum {
+    STOP_GRIPPER = 0x469,
+    START_GRIPPER,
+    SET_PWM,
+    RESET_MCU,
+} CAN_RECEIVE_ID;
+
+typedef enum {
+    I2C_SET_PWM,
+    I2C_STOP_GRIPPER,
+    I2C_START_GRIPPER,
+    I2C_RESET_MCU,
+} I2C_STARTBYTE_ID;
+
+typedef enum {
+    SERVO_1,
+    SERVO_2,
+    SERVO_3,
+} SERVO_ADC_PINS;
+
+void system_init(void);
+
+static inline void enable_servos() {
+    PORT_REGS->GROUP[0].PORT_OUTSET = (1 << 0) | (1 << 27) | (1 << 28);
+}
+static inline void disable_servos() {
+    PORT_REGS->GROUP[0].PORT_OUTCLR = (1 << 0) | (1 << 27) | (1 << 28);
+}
 
 
-void NVMCTRL_Initialize(void);
-
-void PIN_Initialize(void);
-
-void EVSYS_Initialize(void);
-
-
-#ifdef	__cplusplus
+#ifdef __cplusplus
 }
 #endif
 
-#endif	/* SYSTEM_INIT_H */
-
+#endif /* SYSTEM_INIT_H */
